Moved sphere.py tools into structopt/tools
