+++ structopt/tools/__init__.py
@@ -1,2 +1,5 @@
 from . import lammps
 from .parallel import root, single_core, parallel, allgather, parse_MPMD_cores_per_structure, get_rank, get_size
+from .random_three_vector import random_three_vector
+from .get_avg_radii import get_avg_radii
+from .get_particle_radius import get_particle_radius
